reformatting and comments
